requirements.txt
asgiref==3.8.1
click==8.1.7
dj-database-url==2.1.0
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
geographiclib==2.0
geopy==2.4.1
numpy==2.1.2
pillow==10.4.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
